The assembler no creates a definition file (with file-extension .def) which contains all labels of the assembler source in form of constant definitions (.EQU). This definition file should be included in every QNICE-program (together with sysdef.asm) as it defines the call- interface to the monitor (or whatever program source has been compiled).
The Mandelbrotset-program has been adapted accordingly and does now
rely completey on the monitor's IO-routines (thus it is - at least as
of now - not possible to run it "split-screen" with text-output
going to the serial line and "graphics"-output to the VGA-monitor -
the advantage of this is that it now (again) works on the real
hardware as well as on the emulator).

A simple test-program puts.asm has been added which served as a test-
bed for the generation and usage of definition files.