Make assembler script work with GCC 7.5.0
When running this script using GCC 7.5.0, the compiler gives the warning 

```
gcc: warning: monitor.asm: linker input file unused because linking not done
```

and the script produces no output, causing the build to fail.

This is because the gcc frontend examines the filename "monitor.asm" and interprets it as an object file to be fed straight into linking. Any file name with no recognized suffix is treated this way.

The proposed change forces the gcc frontend to interpret the input file as a C file, thus invoking the pre-processor as expected.