Optimize register_file to use LUTRAM instead of registers
The previous implementation created a 3D array of bits, which Vivado is
not able to synthesize as memory and instead creates an array of
registers. This patch changes the implementation to a 2D array of bits,
by synthetically combining the bank and register number to a single
index.

Since LUTRAM reads occur combinatorially the timing is the same as using
registers. Therefore no other changes to the design are necessary.

As an added benefit, the synthesis in Vivado completes much faster too.