Bugfix in UART. Part of Issue #128
These bugs are not likely to show up when the baud rate is 115200, but
they were seen when testing with higher baud rates.

In particular, it was seen during testing that when CTS is de-asserted,
the FPGA will still receive a few more characters, before flow control
sets in. Therefore, CTS should be de-asserted before the FIFO is
completely full.

See Issue #125 for more details of the bugs.