made dev_int simulation work again
The dev_int simulation is used as a development testbed for soft- and hardware interrupts. Due to a multitude of changes in the recent time, the simulation did not work any more. Mainly because a lot of new sources were added and auto-tagged as "used in simulation" by the IDE, which was obviously wrong. So the main task here was to remove all the files that are not needed in simulation and to add some newly added constants.