fat32_library.asm: Changed semantics of FAT32$FILE_SEEK
Seeking is now relative to the start of the file versus relative to the current access position. It is not too hard to make this configurable (as in C's fseek routine), but for now, relative to the start of a file is good enough for our purposes. The changed semantics do not hurt: A grep of the whole existing QNICE code base, including M2M revealed that FAT32$FILE_SEEK was always used directly after opening a file, which is identical to the new semantics.